Refactors Testimonials page, organises dirs
